Show block id for errors in a chain
Currently in a zenroom error is shown the path to the file that raised the
error. After the introduction of `zenContent` this piece of information is
meaningless (`zenContent` doesn't have a path). In this case, we can only
show the block id (inside the yml).

We could show the path when `zenFile` is defined and the block id when
`zenContent` is defined, but it would be ambiguous (when the user sees the
contract identifier he doesn't know if it is a path or a block id)

In the end, for the chains, the choice has been to always show the block id
(inside the yml file)